+++ src/main/java/dev_marcelo/maNotes/repository/AnotacoesRepository.java
@@ -3,6 +3,6 @@ package dev_marcelo.maNotes.repository;
 import dev_marcelo.maNotes.entity.Anotacoes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnotacoesRepository extends JpaRepository<Anotacoes,Float> {
+public interface AnotacoesRepository extends JpaRepository<Anotacoes,Long> {
 
 }
