+++ src/main/java/dev_marcelo/maNotes/repository/FundosEDespesasRepository.java
@@ -3,5 +3,5 @@ package dev_marcelo.maNotes.repository;
 import dev_marcelo.maNotes.entity.FundosEDespesas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FundosEDespesasRepository extends JpaRepository<FundosEDespesas,Float> {
+public interface FundosEDespesasRepository extends JpaRepository<FundosEDespesas,Long> {
 }
