+++ src/main/java/dev_marcelo/maNotes/repository/UsuarioRepository.java
@@ -3,6 +3,6 @@ package dev_marcelo.maNotes.repository;
 import dev_marcelo.maNotes.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRepository extends JpaRepository<Usuario,Float> {
+public interface UsuarioRepository extends JpaRepository<Usuario,Long> {
     Usuario findByLogin(String username);
 }
